Fix: n+1 problem for count

+++ src/main/java/dev/lukaszmichalak/regionalproducts/count/dto/CountDto.java
@@ -2,6 +2,6 @@ package dev.lukaszmichalak.regionalproducts.count.dto;
 
 public record CountDto(
         String voivodeshipName,
-        int productCount
+        long productCount
 ) {
 }
